Added componentless routes
