+++ generators/client/templates/angular/src/main/webapp/app/admin/_admin.state.ts
@@ -1,5 +1,49 @@
-export const adminState = {
-    name: 'admin',
-    abstract: true,
-    parent: 'app'
-};
+import { Routes, CanActivate } from '@angular/router';
+
+import {
+    auditsRoute,
+    configurationRoute,
+    docsRoute,
+    healthRoute,
+    logsRoute,
+    metricsRoute,
+    <%_ if (applicationType === 'gateway') { _%>
+    gatewayState,
+    <%_ } _%>
+    <%_ if (websocket === 'spring-websocket') { _%>
+    trackerState,
+    <%_ } _%>    
+    <%_ if (!skipUserManagement) { _%>
+    userMgmtRoute
+    <%_ } _%>
+} from './';
+
+import { UserRouteAccessService } from '../shared';
+
+let ADMIN_ROUTES = [
+    auditsRoute,
+    configurationRoute,
+    docsRoute,
+    healthRoute,
+    logsRoute,    
+    <%_ if (applicationType === 'gateway') { _%>
+    gatewayRoute,
+    <%_ } _%>
+    <%_ if (websocket === 'spring-websocket') { _%>
+    trackerRoute,
+    <%_ } _%>
+    <%_ if (!skipUserManagement) { _%>
+    ...userMgmtRoute,
+    <%_ } _%>
+    metricsRoute
+];
+
+
+export const adminState: Routes = [{
+    path: '',
+    data: {
+        authorities: ['ROLE_ADMIN']
+    },
+    canActivate: [UserRouteAccessService],
+    children: ADMIN_ROUTES
+}];
