+++ generators/client/templates/angular/src/main/webapp/app/admin/audits/_audits.state.ts
@@ -7,8 +7,6 @@ export const auditsRoute: Route = {
   path: 'audits',
   component: AuditsComponent,
   data: {
-    authorities: ['ROLE_ADMIN'],
     pageTitle: 'audits.title'
-  },
-  canActivate: [UserRouteAccessService]
+  }
 };
