+++ generators/client/templates/angular/src/main/webapp/app/admin/docs/_docs.state.ts
@@ -7,8 +7,6 @@ export const docsRoute: Route = {
   path: 'docs',
   component: <%=jhiPrefixCapitalized%>DocsComponent,
   data: {
-    authorities: ['ROLE_ADMIN'],
     pageTitle: 'global.menu.admin.apidocs'
-  },
-  canActivate: [UserRouteAccessService]
+  }
 };
