+++ generators/client/templates/angular/src/main/webapp/app/admin/health/_health.state.ts
@@ -7,8 +7,6 @@ export const healthRoute: Route = {
   path: 'jhi-health',
   component: <%=jhiPrefixCapitalized%>HealthCheckComponent,
   data: {
-    authorities: ['ROLE_ADMIN'],
     pageTitle: 'health.title'
-  },
-  canActivate: [UserRouteAccessService]
+  }
 };
