+++ generators/client/templates/angular/src/main/webapp/app/admin/logs/_logs.state.ts
@@ -7,8 +7,6 @@ export const logsRoute: Route = {
   path: 'logs',
   component: LogsComponent,
   data: {
-    authorities: ['ROLE_ADMIN'],
     pageTitle: 'logs.title'
-  },
-  canActivate: [UserRouteAccessService]
+  }
 };
