+++ generators/client/templates/angular/src/main/webapp/app/admin/metrics/_metrics.state.ts
@@ -7,8 +7,6 @@ export const metricsRoute: Route = {
   path: 'jhi-metrics',
   component: <%=jhiPrefixCapitalized%>MetricsMonitoringComponent,
   data: {
-    authorities: ['ROLE_ADMIN'],
     pageTitle: 'metrics.title'
-  },
-  canActivate: [UserRouteAccessService]
+  }
 };
